Extract out Preflight details generation
